move ThemeProvider to index.tsx, use dark theme by default on all pages
